# pkg/debuginfo: Cache ELF hashes in ensureUploaded