I've added caching of ELF hashes as it was recommended in the TODO comment.

```
$ go test -benchmem -run=. -bench ^BenchmarkEnsureUploadedInitiateUploadError$ ./pkg/debuginfo/
# new
BenchmarkEnsureUploadedInitiateUploadError-2   	     901	   1415440 ns/op	  422646 B/op	     613 allocs/op
# old
BenchmarkEnsureUploadedInitiateUploadError-2   	     888	   1450389 ns/op	  455509 B/op	     618 allocs/op
```