@javierhonduco 2023-01-16
I am unsure if the added complexity is worth it given that the performance improvement seems small 0.025% improvement. As Kemal says, if this is a very hot code path, it might be worth taking a deeper look and perhaps redesign this part of the codebase