@asubiotto 2023-01-16
drive-by comment: @marselester, you can use the [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) tool to compare benchmark output `$ benchstat old.txt new.txt`. It's a little easier to interpret since it does the performance improvement math for you.